rfc : Minimally packed version of message. Compress all flags into 1 atomic 16bit field.
sizeof(Message) still reports 48, if I look at the raw memory there is a 8 byte field at the very beginning of
the Message class, this is in debug so it is possible -O3 removes that alltogether. Maybe vtable ? since this is base class.
In any case, further compressing all the flags (color/atomic etc...) into 1 gives a very slight boost in single core wrt last commit, more
testing is required.